Update bot.js
Just saved this from my dying laptop. Changed it to use 1 regex instead of 2. Instead of checking for the formatting with another regex, I decided to have it send an error message if the array wasn't the right size for the particular skill check. Now all that's left is to update this from discord.io to discord.js. I already had to create a new token, but a lot of the variables defined in the first 25 lines need to be changed over to the new API. I'll see how to do this in my free time.