Update README.md
Updated the README file to account for the fact that the old code was bad. The new code still needs to be fixed but the programming logic should remain the same for the statements within the switch case structure. I hope to have this back up and running in my spare time.